refactor: WebRTC 연결 리팩토링
