--- src/pages/Consulting/Teller/style.scss
@@ -1,3 +1,0 @@
-.serviceContainer {
-    height: 100%;
-}
