+++ src/pages/Customer/Consulting/index.scss
@@ -5,6 +5,6 @@
 }
 
 #consultLeftSection {
-    width: 70%;
+    width: calc(100% - 300px);
     margin-right: 5%;
 }
