+++ src/pages/Customer/explaining_customer.jsx
