+++ src/pages/Customer/explaining_customer.scss
