+++ src/pages/Customer/loading_customer.scss
